[REFACTOR] #22 - GithubApiRepository 삭제 및 TokenRepository로 기능 이관
